refactor into a cleaner visitor pattern thing and add some comments
