+++ src/package/visitors/tables/get-ttl.js
@@ -1,0 +1,9 @@
+module.exports = function getTTL(attr) {
+  var found = false
+  Object.keys(attr).forEach(k=> {
+    if (attr[k] === 'TTL') {
+      found = k
+    }
+  })
+  return found
+}
